+++ src/main/java/de/bauersoft/data/providers/AllergenDataProvider.java
@@ -1,20 +1,18 @@
 package de.bauersoft.data.providers;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Stream;
-
-import org.springframework.stereotype.Service;
-
 import com.vaadin.flow.data.provider.ConfigurableFilterDataProvider;
 import com.vaadin.flow.data.provider.DataChangeEvent;
 import com.vaadin.flow.data.provider.DataProviderListener;
 import com.vaadin.flow.data.provider.Query;
 import com.vaadin.flow.shared.Registration;
-
-import de.bauersoft.data.entities.Allergen;
+import de.bauersoft.data.entities.allergen.Allergen;
 import de.bauersoft.data.filters.SerializableFilter;
 import de.bauersoft.services.AllergenService;
+import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Stream;
 
 @Service
 public class AllergenDataProvider implements ConfigurableFilterDataProvider<Allergen, Void, List<SerializableFilter<Allergen,?>>> {
@@ -33,7 +31,7 @@ public class AllergenDataProvider implements ConfigurableFilterDataProvider<Alle
 
 	@Override
 	public int size(Query<Allergen, Void> query) {
-		return this.service.count(filter);
+		return (int) this.service.count(filter);
 	}
 
 	@Override
