+++ src/main/java/de/bauersoft/data/providers/ComponentDataProvider.java
@@ -1,20 +1,18 @@
 package de.bauersoft.data.providers;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Stream;
-
-import org.springframework.stereotype.Service;
-
 import com.vaadin.flow.data.provider.ConfigurableFilterDataProvider;
 import com.vaadin.flow.data.provider.DataChangeEvent;
 import com.vaadin.flow.data.provider.DataProviderListener;
 import com.vaadin.flow.data.provider.Query;
 import com.vaadin.flow.shared.Registration;
-
-import de.bauersoft.data.entities.Component;
+import de.bauersoft.data.entities.component.Component;
 import de.bauersoft.data.filters.SerializableFilter;
 import de.bauersoft.services.ComponentService;
+import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Stream;
 
 @Service
 public class ComponentDataProvider implements ConfigurableFilterDataProvider<Component, Void, List<SerializableFilter<Component,?>>> {
@@ -33,7 +31,7 @@ public class ComponentDataProvider implements ConfigurableFilterDataProvider<Com
 
 	@Override
 	public int size(Query<Component, Void> query) {
-		return this.service.count(filter);
+		return (int) this.service.count(filter);
 	}
 
 	@Override
