+++ src/main/java/de/bauersoft/data/providers/CourseDataProvider.java
@@ -1,20 +1,18 @@
 package de.bauersoft.data.providers;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Stream;
-
-import org.springframework.stereotype.Service;
-
 import com.vaadin.flow.data.provider.ConfigurableFilterDataProvider;
 import com.vaadin.flow.data.provider.DataChangeEvent;
 import com.vaadin.flow.data.provider.DataProviderListener;
 import com.vaadin.flow.data.provider.Query;
 import com.vaadin.flow.shared.Registration;
-
-import de.bauersoft.data.entities.Course;
+import de.bauersoft.data.entities.course.Course;
 import de.bauersoft.data.filters.SerializableFilter;
 import de.bauersoft.services.CourseService;
+import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Stream;
 
 @Service
 public class CourseDataProvider  implements ConfigurableFilterDataProvider<Course, Void, List<SerializableFilter<Course,?>>> {
@@ -38,7 +36,7 @@ public class CourseDataProvider  implements ConfigurableFilterDataProvider<Cours
 	
 	@Override
 	public int size(Query<Course, Void> query) {
-		return this.service.count(filter);
+		return (int) this.service.count(filter);
 	}
 
 	@Override
