+++ src/main/java/de/bauersoft/data/providers/MenuDataProvider.java
@@ -5,7 +5,7 @@ import com.vaadin.flow.data.provider.DataChangeEvent;
 import com.vaadin.flow.data.provider.DataProviderListener;
 import com.vaadin.flow.data.provider.Query;
 import com.vaadin.flow.shared.Registration;
-import de.bauersoft.data.entities.Menu;
+import de.bauersoft.data.entities.menu.Menu;
 import de.bauersoft.data.filters.SerializableFilter;
 import de.bauersoft.services.MenuService;
 import org.springframework.stereotype.Service;
@@ -47,7 +47,7 @@ public class MenuDataProvider implements ConfigurableFilterDataProvider<Menu, Vo
     @Override
     public int size(Query<Menu, Void> query)
     {
-        return this.service.count(filter);
+        return (int) this.service.count(filter);
     }
 
     @Override
