+++ src/main/java/de/bauersoft/data/providers/RecipeDataProvider.java
@@ -1,21 +1,15 @@
 package de.bauersoft.data.providers;
 
+import com.vaadin.flow.data.provider.*;
+import com.vaadin.flow.shared.Registration;
+import de.bauersoft.data.entities.recipe.Recipe;
+import de.bauersoft.data.filters.SerializableFilter;
+import de.bauersoft.services.RecipeService;
+import org.springframework.stereotype.Service;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
-
-import org.springframework.stereotype.Service;
-
-import com.vaadin.flow.data.provider.ConfigurableFilterDataProvider;
-import com.vaadin.flow.data.provider.DataChangeEvent;
-import com.vaadin.flow.data.provider.DataProvider;
-import com.vaadin.flow.data.provider.DataProviderListener;
-import com.vaadin.flow.data.provider.Query;
-import com.vaadin.flow.shared.Registration;
-
-import de.bauersoft.data.entities.Recipe;
-import de.bauersoft.data.filters.SerializableFilter;
-import de.bauersoft.services.RecipeService;
 
 @Service
 public class RecipeDataProvider implements ConfigurableFilterDataProvider<Recipe, Void, List<SerializableFilter<Recipe,?>>>, DataProvider<Recipe, Void> {
@@ -34,7 +28,7 @@ public class RecipeDataProvider implements ConfigurableFilterDataProvider<Recipe
 
 	@Override
 	public int size(Query<Recipe, Void> query) {
-		return this.service.count(filter);
+		return (int) this.service.count(filter);
 	}
 
 	@Override
